Move the patterns file into the validators package
