+++ validators/patterns.go
@@ -1,4 +1,4 @@
-package utils
+package validators
 
 const (
 	LowerCasePattern   = `.*[[:lower:]]+.*[[:lower:]]+.*`
